ordinary numbers

+++ main.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
-int countOrdinaryNumbers(int n) {
-    int count = 0;
+long long countOrdinaryNumbers(long long n) {
+    long long count = 0;
     for (int digit = 1; digit <= 9; ++digit) {
-        int ordinaryNumber = digit;
+        long long ordinaryNumber = digit;
         while (ordinaryNumber <= n) {
             count++;
             ordinaryNumber = ordinaryNumber * 10 + digit;
@@ -14,7 +14,7 @@ int main() {
     int t;
     std::cin >> t;
     while (t--) {
-        int n;
+        long long n;
         std::cin >> n;
         std::cout << countOrdinaryNumbers(n) << std::endl;
     }
